improve triangulation
The earcut4j library sometimes creates incorrect triangulation results which may lead to small rendering artefacts (e.g., triangles are created that go beyond the polygon). The JTS triangulation seems to produce better results, but is a bit slower and is not always able to compute the triangles. The changes first try the ConstrainedDelaunayTriangulator from JTS and if that fails, earcut4j is used.

In addition, the orientation (ccw or cw) was not detected/applied correctly.