remove GML pass through encoder
This commit removes the "pass through" option for the GML format. The reason is that upgrading the pass-through-logic for the new onGeometry would have been complex and the pass-through-option is only used for WFS feature providers and GML output, a use case that is by now exotic and unnecessary.

Is there still a use case for the pass-through option for some other combination of feature provider and output format?