remove TODO
After more thoughts I think that triangle meshes as generated by this class are quite different from the other geometry types. The data structures do not directly represent the geometry, but the specific glTF representation. So, at least for now it is better to keep the class separate from the other geometry types.